Add alternate version of "Sonic the Hedgehog - Water Warning (UCantBreathe2This)"
As far as I can tell, this track was only available in 160kbps CBR (checked both OCRemix on Internet Archive, and the VGMix Archive).  I am unsure where this VBR version came from, but believe it may be a later re-encode.  Since I'm not ready to rule it out just yet, the 160kbps version is now the standard, while the VBR one is kept as an "alt"ernate.